+++ src/components/Items/normativas/styleNorm.css
@@ -21,7 +21,7 @@
 
 .icon {
   float: right;
-  margin-left: 10rem;
+  padding: 0.3rem;
 }
 .download-icon {
   font-size: 1.5rem;
